+++ internal/genre/service/service.go
@@ -4,7 +4,7 @@ import (
 	"net/http"
 	"strconv"
 
-	"Demo/internal/models"
+	"Demo/internal/genre/models"
 
 	"github.com/gin-gonic/gin"
 	"gorm.io/gorm"
@@ -18,7 +18,7 @@ func GetGenreById(db *gorm.DB) gin.HandlerFunc {
 			return
 		}
 
-		var genre []models.Genre
+		var genre []genreModels.Genre
 		result := db.First(&genre, id)
 		if result.Error != nil {
 			c.JSON(http.StatusNotFound, gin.H{"error": "Genre not found"})
@@ -30,7 +30,7 @@ func GetGenreById(db *gorm.DB) gin.HandlerFunc {
 
 func CreateGenre(db *gorm.DB) gin.HandlerFunc {
 	return func(c *gin.Context) {
-		var genre models.Genre
+		var genre genreModels.Genre
 		err := c.ShouldBindJSON(&genre)
 		if err != nil {
 			c.JSON(http.StatusBadRequest, gin.H{"error": "Invalid request payload"})
@@ -55,7 +55,7 @@ func UpdateGenre(db *gorm.DB) gin.HandlerFunc {
 			return
 		}
 
-		var genre models.Genre
+		var genre genreModels.Genre
 		result := db.First(&genre, id)
 		if result.Error != nil {
 			c.JSON(http.StatusNotFound, gin.H{"error": "Genre not found"})
@@ -86,7 +86,7 @@ func DeleteGenre(db *gorm.DB) gin.HandlerFunc {
 			return
 		}
 
-		var genre models.Genre
+		var genre genreModels.Genre
 		result := db.First(&genre, id)
 		if result.Error != nil {
 			c.JSON(http.StatusNotFound, gin.H{"error": "Genre not found"})
